Moving trace testing code to separate source file
